Improve RTL support
For all the layout elements having margin, padding,... to right or left,
releavnt start/end property is added so all the elements are supposed to
be displayed properly in RTL mode.

There are still minor issues related to some icons such as play, ff and
rev. This issue must be addressed separately in future.