Fixed playback speed issues when switching to next media
Fixes some issues related to the playback speed when playback was
completed or playback was paused and the app went away for awhile.

Now a media item which is "in progress" will remember its playback
speed so that it can be restored properly. The per-feed-media speed is
cleared once playback finishes, either by reaching the end or by the
user starting to play something different.