Reduced probability for closing player accidentally
If the description is really short, the WebView shrinks. When touching
below the WebView, one needs to slide pretty precisely to swipe horizontally.
If there is a little bit of slope, the player closes (completely, without following
the finger). This change resizes the WebView, so that one can not touch below it.