Remove IF NOT EXISTS from index create statements
`onCreate()` should only ever be called once, where the tables and
indices are created. Any other changes need to happen via `onUpgrade()`,
so we can safely remove `IF NOT EXISTS`.

When importing any database into the app (e.g. if it gets corrupted and
is recreated with `sqlite3 old-database.db ".dump"`), it first needs to
be set back to the correct version. This can be done in sqlite with:

    $ sqlite3 old-database.db "PRAGMA user_version"
    <old-db-version>
    $ sqlite3 new-database.db "PRAGMA user_version = <old-db-version>"

For more context, see this PR:
https://github.com/AntennaPod/AntennaPod/pull/4585