Added new cleanup option: when not favorited
This is another way of solving #2077.

The root issue is that queued episodes are never auto-deleted
currently which means that if you automatically add episodes to the
queue you will eventually end up with AntennaPod refusing to auto
download more episodes because the cache is full and it can't make
space.

This option will only refuse to delete favorited items. Otherwise it
will simply delete the oldest episodes.