Avoid stuck download notifications
Cancel the periodic updater before calling stopForeground so
that the notification can't come back after it has been closed.

Remove the workaround for API 30 slow notification closing because
it can cause a stuck notification. However, in order to avoid misleading
text being left on the notification if it closes slowly, ensure that
the "0 downloads" text is written to the notification when shutting down.

Fixes #4556